+++ .config/nvim/lua/config/init.lua
@@ -1,3 +1,2 @@
 require'config.galaxyline'
-require'config.fzf'
 require'config.colorscheme'
