--- .config/nvim/lua/lsp_lua/rust_analyzer.lua
@@ -1,19 +1,0 @@
-return {
-    on_attach=function(client, bufnr)
-        require'completion'.on_attach(client)
-    end,
-    setttings = {
-        ["rust-analyzer"] = {
-            assist = {
-                importGranularity = "module",
-                importPrefix = "by_self",
-            },
-            cargo = {
-                loadOutDirsFromCheck = true
-            },
-            procMacro = {
-                enable = true
-            },
-        }
-    }
-}
