--- .config/nvim/lua/lsp_lua/texlab.lua
@@ -1,17 +1,0 @@
-return {
-  settings = {
-        texlab = {
-            build = {
-                executable = "tectonic";
-                args = {
-                    "%f",
-                    "--synctex",
-                    "--keep-logs",
-                    "--keep-intermediates",
-                };
-                onSave = true;
-            };
-        };
-    },
-    on_attach = function() end,
-}
